services/speech-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai-whisper==20231117
python-multipart==0.0.6
websockets==12.0
numpy>=1.24.0
torch>=2.0.0
